fix: Fix the remaining warnings

+++ src/layer/riscv/celu_riscv_zfh.cpp
@@ -34,10 +34,10 @@ int CELU_riscv::forward_inplace_fp16s(Mat& bottom_top_blob, const Option& opt) c
             vfloat16m8_t _p = __riscv_vle16_v_f16m8(ptr, vl);
             vbool2_t _mask = __riscv_vmfgt_vf_f16m8_b2(_p, (__fp16)0.f, vl);
 
-            vfloat16m8_t _q = __riscv_vfdiv_vf_f16m8(_p, alpha, vl);
+            vfloat16m8_t _q = __riscv_vfdiv_vf_f16m8(_p, (__fp16)alpha, vl);
             _q = exp_ps(_q, vl);
             _q = __riscv_vfsub_vf_f16m8(_q, (__fp16)1.f, vl);
-            _q = __riscv_vfmul_vf_f16m8(_q, alpha, vl);
+            _q = __riscv_vfmul_vf_f16m8(_q, (__fp16)alpha, vl);
 
             vfloat16m8_t _res = __riscv_vmerge_vvm_f16m8(_q, _p, _mask, vl);
             __riscv_vse16_v_f16m8(ptr, _res, vl);
